[definite-init] Remove dead code.
collectRetainCountInfo bails early if our memory instruction is a
mark_uninitialized... but our instruction is always a mark_uninitialized... so
this code is dead.

The origin of this code is from the flattening of the control flow in DI that
was necessary to be done to "extract" predictable mem opts from it.

rdar://40332620